Log in to
Apparently "log in" is a verb (I need to log in) and "login" is a noun (Go to login, don't give out your login) or adjective  (the login screen). This seems kind of silly, but I guess it does avoid the possibility of creating the obviously wrong constructions like, "This happens when user logins to the account" or "he needs to be loginned to see the permissions". I from now on am going to adopt this as a hard rule so that we have a consistent style.

Also, its " log in to", not "log in into" .